add files from aws lambda
